feat: move logout button to settings page
Remove logout button from all view headers (home, recipes, grocery-lists)
and add it to the Settings page at the bottom in a "Session" section.
Delete the now-unused LogoutButton component.
